Initial commit of mindrender package skeleton
This commit contins a skeleton setup for creating and
uploading the mindrender package.  The package skeleton
supports building the package, installing it, testing
and github workflow testing of the package.  We haven't
yet added in code yet, this is just the build and
test framework.  This framework was cribbed from the
following tutorial:

https://circleci.com/blog/publishing-a-python-package