Commit version 0.0.5, add skeleton plot and render functions
- Added skeleton of plot functions meant to plot a single frame and/or plot a
  motion track as a figure at some point in time
- Added skeleton of render functions.  These return an animation object that can be
  saved as a movie file or other video type.  NOTE: not sure if the animation
  uses ffmpeg on back end or that is not a dependency anymore.

NOTE: these are a bit bare bones, need to add documentaiton and better
tests.

TODO: Should add an issue to clean up the UI, can make much simpler by
inferring joint_graph and joint_names from time_df.  That is probably
all we need to have the user give for the plot and render functions.

TODO: also we will have problems with parameter/values for
titles, time interval between frames, etc.  If we really need this to
generalize, would probably need to add in mechanisms to specify this
in config files.