Version 0.0.5 commit, add plot and render functions (#3)
* Version 0.0.4 commit, preliminary data loading and parsing

- Added basic function to load time series and joint graph information.
- These function currently perform some basic parsing and error checking.
- Added test of goot parses of these files, still should add additional
  testing of error conditions on bad file inputs.

* Fix pandas dependency in ci

Add in install of pandas library, used as module in the
load data util functions.

* Commit version 0.0.5, add skeleton plot and render functions

- Added skeleton of plot functions meant to plot a single frame and/or plot a
  motion track as a figure at some point in time
- Added skeleton of render functions.  These return an animation object that can be
  saved as a movie file or other video type.  NOTE: not sure if the animation
  uses ffmpeg on back end or that is not a dependency anymore.

NOTE: these are a bit bare bones, need to add documentaiton and better
tests.

TODO: Should add an issue to clean up the UI, can make much simpler by
inferring joint_graph and joint_names from time_df.  That is probably
all we need to have the user give for the plot and render functions.

TODO: also we will have problems with parameter/values for
titles, time interval between frames, etc.  If we really need this to
generalize, would probably need to add in mechanisms to specify this
in config files.

* Fix ci issue, don't need to import mpl_toolkit

We were importing the mpl_toolkit because of need
of 3d axis, but this is legacy code.  This
dependency seem to be no longer needed, and can just
import matplotlib which contains matplotlib.animation
as well as support for 3d extensions on axis.

* Fix ci issue missing matplotlib dependency

Should have also thought of this in previous commit.
We have added matplotlib as a dependency in the
plot and render submodules, so need to add it
into the ci environment.

* Matplotlib animation does depend on ffmpeg or other back end

So matplotlib.animation does still render on backend using
ffmpeg or other renderer of the frames.  For now we will
disable the render animation tests and come back to this
when we figure out how to do some better testing.

* Fix indentation syntax error

Not always a good idea to use the github merge tool.
Accidentally misindented a line when resolving
conflicts.  Now fixed.